src/fetcher/pyproject.toml
[tool.poetry]
name = "fetcher"
version = "0.1.0"
description = ""
authors = ["Taina Hilliaho <taina.hilliaho@helsinki.fi>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
pymongo = "^4.12.1"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
